requirements.txt
Django==3.2.19
djangorestframework==3.13.1
python-decouple==3.8
psycopg2-binary==2.9.6
dj-database-url==2.0.0
drf-yasg2==1.19.4
django-keycloak-auth==0.9.6
django-cors-headers==4.0.0
pytz==2023.3
djangorestframework-simplejwt==5.2.2
django-debug-toolbar==3.2.2
ipython
django-extensions==3.1.3
django-import-export==2.6.0
pandas==2.1.0
Unidecode==1.3.7
gunicorn==20.1.0
